requirements.txt
fastapi
uvicorn[standard]
aiofiles
aiosqlite
python-multipart
sqlalchemy[asyncio]
asyncpg
SQLAlchemy>=1.4
python-multipart
python-dotenv
requests
psycopg2
